+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 * Added support for HVAC
 * Added support for excluding devices types from the compilation using defines (see `src/ArduinoHADefines.h`)
 * Added support for setting icon in HASwitch and HASensor
+* Added support for setting retain flag in HASwitch
+
 
 **Updates:**
 * Optimized codebase and logic in all devices types